Remove stripping of leading task name in paths
The assumption that we should strip the task name path prefix
is no longer correct if we want to be able to browse the
allocation without a task. This changes the task filesystem
acceptance test files to all be nested within a task name
directory, which mirrors how it works in the real setup.