Add specificity to exec allocation URL generation
Thanks to @notnoop for this UX improvement suggestion.
The allocation’s task group is always known, so it
might as well be preselected in the sidebar when the
exec window opens. Also, if the task group only has
one task, might as well preselect it too.

This is perhaps at or beyond the limit of where it makes
sense to use non-Ember Data pseudo-models in the tests 😶