Fix layout of title component elements
The spacing has been broken for job types that use this yield
(parameterised and periodic) since I added the exec button
to this template. The grouping here may be subpar; another
approach would be to yield contextual components (in the
absence of named blocks) so invocations can specify which
grouping to add to.