docs: remove mention of custom schedulers
Not sure if this was meant to imply adding more schedulers to Nomad is
easy, or that we plan on adding pluggable schedulers. Either way,
neither of those statements is really true unless you really stretch the
definitions of "easy" and "plan".

So remove this sentence as I can't imagine it does anything other than
confuse people.