Fix label failures
This rule is described here:
https://dequeuniversity.com/rules/axe/3.5/label

While adding a label for the token input field, I noticed
that its placeholder produces a bad VoiceOver experience,
which I will address separately.

I removed the action to focus the stepper input when
clicking the label as I believe connecting the label
and input with an id is equivalent.

This includes an override for ivy-codemirror’s codemirror
dependency to take advantage of the screenReaderLabel
configuration option in 5.53; thanks to @adityatoshniwal
for that addition in codemirror/CodeMirror#6197.