Add preliminary component auditing
I decided to add these to existing tests vs create new ones
because audits will need to happen for the various states that
components can be in and it seems wasteful to need to reproduce
all the iterations separately vs use what already exists. I also
decided to fold the assertion into the audit wrapper so it doesn’t
need to be duplicated everywhere.