Add linting to ensure component audits happen
This doesn’t catch most of the component tests because of the
directory structure… sadly it seems like many files should be
moved. I could perhaps instead use an elaborate filter to
ensure integration/util is excluded but everythig else is
included but considering that `ember g component-test …`
generates the file in integration/components, it seems best
to move everything there.