Add colour hacks to satisfy Axe color-contrast
This is a brutish mass changing of colour values. It’s mostly
darkening or lightening existing colours to increase contrast,
with no regard to æsthetics.

Since this isn’t meant to actually be merged, there are still
some failures, as it’s more of a general survey and the
randomisation of test data makes it difficult to reproduce
errors. If we wanted to start actually exercising the
color-contrast rule, it would perhaps help to have some
component tests deliberately iterate through the various
colour combinations. boxed-section was a particular trouble
spot and I ultimately removed some audits because it was
a timesink.