Allow operators to configure cleanup_on_shutdown for clients
In dev mode the client will clean up allocation resources on shutdown, but there is no good way to get this in non-dev mode scenarios, and there are use cases for it - such as running a temporary client nodes.

This PR simply exposes a config option and deletes the DevMode test flag prefering to use it instead - defaulting it to true in dev mode.