Prune Events on insert within same transaction
Prior to this change, we attempted to prune events within their own
transaction. This was uncoordinated with the FSM and caused competing
FSM requests and event evictions to lock up go-memdb when they both
tried to acquire a write transaction.

This change fixes this by inserting events and deleting old events to
constrain the table by the configured size all within the main FSM
transaction that generated the events.