Fix infinite retry if err is "no leader"
Moving the timeout calculation to the HasTimedOut method effectively
removed it from the code path where either the args aren't implementing
RPCInfo or if the error is "No cluster leader". This fix brings the
timeout calculation before everything else and works as originally
designed even if the args don't implement the RPCInfo interface.