added setVersion
setVersion adds a header X-Nomad-Version to a wrapper function. There is
a wrapper function in http.go that takes a normal implementation for an
endpoint, and adds other code that should be there. As an example, let's
say that an endpoint wants to search for all the jobs running. The
function would implement the logic to look for jobs. Then you take that
function, pass it to the wrapper, and the wrapper will add all the
headers that should be on that response.

It's nice because all of the logic that you'd have to add version
headers can get put in one place